Remove /thread/ route; consolidate onto /p/
The /thread/[threadId]/, /thread/[threadId]/info/, and
/thread/[threadId]/edit-title/ routes are gone. Their pages move under
/p/[shortId]/ and the empty thread placeholder (was /thread/new/) becomes
the bare /p/ route. The collapse-card branch in the URL replaceState is
also dropped — collapsing leaves the URL on the just-collapsed poll, so
"thread" becomes a view of the poll page rather than its own route.
